game-logic: added cord_to_cell matrix

+++ src/core/cell.rs
@@ -1,7 +1,7 @@
 #[derive(Default)]
 pub struct Cell {
-    pub x: i64,
-    pub y: i64,
+    pub x: usize,
+    pub y: usize,
     // ways is a vector of indexes of the cells to which we can go from the current one
     pub next_cells: Vec<usize>,
     pub ghost_presence: bool,
